requirements.txt
assimp-py==1.0.6
numpy==1.23.5
pygame==2.1.2
PyOpenGL==3.1.6
typer==0.7.0